+++ src/providers/twitch/TwitchIrcServer.cpp
@@ -15,6 +15,7 @@
 #include "providers/twitch/TwitchAccount.hpp"
 #include "providers/twitch/TwitchChannel.hpp"
 #include "providers/twitch/TwitchHelpers.hpp"
+#include "singletons/Settings.hpp"
 #include "util/PostToThread.hpp"
 
 #include <QMetaEnum>
@@ -351,7 +352,11 @@ void TwitchIrcServer::onMessageSendRequested(TwitchChannel *channel,
                                 ? this->lastMessageMod_
                                 : this->lastMessagePleb_;
         size_t maxMessageCount = channel->hasHighRateLimit() ? 99 : 19;
-        auto minMessageOffset = (channel->hasHighRateLimit() ? 100ms : 1100ms);
+        const auto lowRateLimit = getSettings()->twitchLowRateLimitDelay;
+        const auto highRateLimit = getSettings()->twitchHighRateLimitDelay;
+        auto minMessageOffset = (channel->hasHighRateLimit()
+                                     ? std::chrono::milliseconds(highRateLimit)
+                                     : std::chrono::milliseconds(lowRateLimit));
 
         auto now = std::chrono::steady_clock::now();